+++ R/cfa_fit_plot.R
@@ -1,6 +1,7 @@
 #' @title Fit and plot CFA simultaneously
 #'
-#' @description Print and save CFA fit, as well as plot CFA factor loadings, simultaneously.
+#' @description Prints and saves CFA fit, as well as plots CFA factor loadings,
+#'              simultaneously.
 #'
 #' @param model CFA model to fit.
 #' @param data Data set on which to fit the CFA model.
@@ -9,16 +10,21 @@
 #' @param ... Arguments to be passed to function `lavaan::cfa`.
 #' @param remove.items Optional, if one wants to remove items from the CFA model
 #'                     without having to redefine it completely again.
+#' @param print Logical, whether to print model summary to console.
 #' @param save.as.pdf Logical, whether to save as PDF for a high-resolution,
 #'                    scalable vector graphic quality plot. Defaults to
-#'                    saving to the "/model" subfolder of the working directory.
-#'                    If it doesn't exist, it creates it. Then automatically
-#'                    open the created PDF in the default browser. Defaults to false.
-#' @param file.name Optional (when `save.as.pdf` is set to `TRUE`), if one wants
-#'                  something different than the default file name. It saves to
-#'                  pdf per default, so the .pdf extension should not be specified
-#'                  as it will add it automatically.
-#' @keywords CFA, lavaan, plot, fit
+#'                    saving to the "/model" subfolder of the working
+#'                    directory. If it doesn't exist, it creates it. Then
+#'                    automatically open the created PDF in the default
+#'                    browser. Defaults to false.
+#' @param file.name Optional (when `save.as.pdf` is set to `TRUE`), if one
+#'                  wants something different than the default file name.
+#'                  It saves to pdf per default, so the .pdf extension
+#'                  should not be specified as it will add it automatically.
+#' @keywords CFA lavaan plot fit
+#' @return The function returns a `lavaan` fit object. However, it also: prints
+#'         a summary of the `lavaan` fit object to the console, and; prints a
+#'         `lavaanPlot` of the `lavaan` fit object.
 #' @export
 #' @examples
 #' (latent <- list(visual = paste0("x", 1:3),
@@ -33,18 +39,13 @@
 #' @import lavaan
 #' @section Illustrations:
 #'
-#' \if{html}{\figure{cfaplot.png}{options: width="400"}}
+#' \if{html}{
+#' \figure{cfaplot.png}{options: width="400"}
+#' }
 
-cfa_fit_plot <- function(model, data, covs = FALSE, estimator = "MLR", remove.items = "",
-                         save.as.pdf = FALSE, file.name, ...){
-  if(missing(file.name) && isTRUE(save.as.pdf)) {
-    prefix <- deparse(substitute(model))
-    time <- gsub(":", ".", Sys.time())
-    file.name <- paste0("models/cfa_", prefix, "_", time)
-    if (!dir.exists("models")) {
-      dir.create("models")
-      }
-  }
+cfa_fit_plot <- function(
+    model, data, covs = FALSE, estimator = "MLR", remove.items = "",
+    print = TRUE, save.as.pdf = FALSE, file.name, ...){
   # Remove requested items
   if(!missing(remove.items)) {
     remove.items0 <- paste0("\\s", remove.items, "\\s")
@@ -60,25 +61,17 @@ cfa_fit_plot <- function(model, data, covs = FALSE, estimator = "MLR", remove.it
   }
   # Fit model
   fit <- lavaan::cfa(model, data = data, estimator = estimator, ...)
-  print(summary(fit, standardized = TRUE, fit.measures = TRUE, rsquare = TRUE))
+  if (isTRUE(print)) {
+    print(summary(fit, standardized = TRUE, fit.measures = TRUE, rsquare = TRUE))
+  }
   # Plot the model
   my.plot <- nice_lavaanPlot(fit, covs = covs)
   # # Save file
-  if(isTRUE(save.as.pdf)) {
+  if(!missing(file.name) && isTRUE(save.as.pdf)) {
     lavaanPlot::embed_plot_pdf(my.plot, paste0(file.name, ".pdf"))
     utils::browseURL(paste0(file.name, ".pdf"))
   } else {
     print(my.plot)
   }
   fit
-}
-
-serialNext <- function(prefix){
-  if(!file.exists(paste0(prefix, ".pdf"))){return(prefix)}
-  i <- 1
-  repeat {
-    f = paste0(prefix, "_", i)
-    if(!file.exists(f)){return(f)}
-    i <- i + 1
-  }
 }