+++ R/lavaan_cov.R
@@ -6,10 +6,13 @@
 #'              `std.all` column.
 #'
 #' @param fit lavaan fit object to extract covariance indices from
-#' @param nice_table Logical, whether to print the table as a `rempsyc::nice_table`
-#'                   as well as print the reference values at the bottom of the table.
+#' @param nice_table Logical, whether to print the table as a
+#'                   `rempsyc::nice_table` as well as print the
+#'                   reference values at the bottom of the table.
 #' @param ... Arguments to be passed to `rempsyc::nice_table`
-#' @keywords lavaan, structural equation modeling, path analysis, CFA
+#' @keywords lavaan structural equation modeling path analysis CFA
+#' @return A dataframe of covariances, including the covaried variables,
+#'         the covariance, and corresponding p-value.
 #' @export
 #' @examples
 #'
@@ -38,20 +41,7 @@ lavaan_cov <- function(fit, nice_table = FALSE, ...) {
   x <- x[c("lhs", "rhs", "std.all", "pvalue")]
   names(x) <- c("Variable 1", "Variable 2", "r", "p")
   if (nice_table) {
-    if (isFALSE(requireNamespace("rempsyc", quietly = TRUE))) {
-      cat("The package `rempsyc` is required for this feature\n",
-          "Would you like to install it?")
-      if (utils::menu(c("Yes", "No")) == 1) {
-        utils::install.packages('rempsyc', repos = c(
-          rempsyc = 'https://rempsyc.r-universe.dev',
-          CRAN = 'https://cloud.r-project.org'))
-      } else (stop(
-        'The `nice_table` feature relies on the `rempsyc` package.
-    You can install it manually with:
-    install.packages("rempsyc", repos = c(
-          rempsyc = "https://rempsyc.r-universe.dev",
-          CRAN = "https://cloud.r-project.org")'))
-    }
+    rlang::check_installed("rempsyc", reason = "for this feature.")
     x <- rempsyc::nice_table(x, ...)
   }
   x