+++ R/nice_fit.R
@@ -4,13 +4,18 @@
 #' optionally includes references values. The reference fit values are
 #' based on Schreiber et al. (2006).
 #'
-#' Schreiber, J. B., Nora, A., Stage, F. K., Barlow, E. A., & King, J. (2006). Reporting structural equation modeling and confirmatory factor analysis results: A review. *The Journal of educational research*, *99*(6), 323-338. https://doi.org/10.3200/JOER.99.6.323-338
-#'
 #' @param ... lavaan model objects to extract fit indices from
-#' @param nice_table Logical, whether to print the table as a `rempsyc::nice_table`
-#'                   as well as print the reference values at the bottom of the table.
-#' @keywords lavaan, structural equation modeling, path analysis, CFA
+#' @param nice_table Logical, whether to print the table as a
+#'                   `rempsyc::nice_table` as well as print the
+#'                   reference values at the bottom of the table.
+#' @keywords lavaan structural equation modeling path analysis CFA
+#' @return A dataframe, representing select fit indices (chi2, df, chi2/df,
+#'         p-value of the chi2 test, CFI, TLI, RMSEA, SRMR, AIC, and BIC).
 #' @export
+#' @references Schreiber, J. B., Nora, A., Stage, F. K., Barlow, E. A., & King,
+#' J. (2006). Reporting structural equation modeling and confirmatory factor
+#' analysis results: A review. *The Journal of educational research*, *99*(6),
+#' 323-338. https://doi.org/10.3200/JOER.99.6.323-338
 #' @examples
 #' (latent <- list(visual = paste0("x", 1:3),
 #'                 textual = paste0("x", 4:6),
@@ -23,45 +28,32 @@
 #' cat(HS.model)
 #'
 #' library(lavaan)
-#' fit <- lavaan(HS.model, data=HolzingerSwineford1939,
-#'               auto.var=TRUE, auto.fix.first=TRUE,
-#'               auto.cov.lv.x=TRUE)
+#' fit <- sem(HS.model, data=HolzingerSwineford1939)
 #' nice_fit(fit)
 
 nice_fit <- function(..., nice_table = FALSE) {
   x <- lapply(list(...), nice_fit_internal)
   df <- do.call(rbind, x)
-  Model <- sapply(match.call(expand.dots = FALSE)$`...`, as.character)
+  Model <- vapply(match.call(expand.dots = FALSE)$`...`, as.character,
+                  FUN.VALUE = "character")
   df <- cbind(Model, df)
   row.names(df) <- NULL
   if (nice_table) {
-    if (isFALSE(requireNamespace("rempsyc", quietly = TRUE))) {
-      cat("The package `rempsyc` is required for this feature\n",
-          "Would you like to install it?")
-      if (utils::menu(c("Yes", "No")) == 1) {
-        utils::install.packages('rempsyc', repos = c(
-          rempsyc = 'https://rempsyc.r-universe.dev',
-          CRAN = 'https://cloud.r-project.org'))
-      } else (stop(
-        'The `nice_table` feature relies on the `rempsyc` package.
-    You can install it manually with:
-    install.packages("rempsyc", repos = c(
-          rempsyc = "https://rempsyc.r-universe.dev",
-          CRAN = "https://cloud.r-project.org")'))
-    }
+    rlang::check_installed("rempsyc", reason = "for this feature.")
     table <- rempsyc::nice_table(df)
-    table <- flextable::add_footer_row(table, values = c(Model = "Ideal Value",
-                                                         chi2 = "\u2014",
-                                                         df = "\u2014",
-                                                         chi2.df = "< 2 or 3",
-                                                         p = "> .05",
-                                                         CFI = "\u2265 .95",
-                                                         TLI = "\u2265 .95",
-                                                         RMSEA = "< .06-.08",
-                                                         SRMR = "\u2264 .08",
-                                                         AIC = "Smaller is better",
-                                                         BIC = "Smaller is better"),
-                                       colwidths = rep(1, length(table$col_keys)))
+    table <- flextable::add_footer_row(table, values = c(
+      Model = "Ideal Value",
+      chi2 = "\u2014",
+      df = "\u2014",
+      chi2.df = "< 2 or 3",
+      p = "> .05",
+      CFI = "\u2265 .95",
+      TLI = "\u2265 .95",
+      RMSEA = "< .06-.08",
+      SRMR = "\u2264 .08",
+      AIC = "Smaller is better",
+      BIC = "Smaller is better"),
+      colwidths = rep(1, length(table$col_keys)))
     table <- flextable::bold(table, part = "footer")
     table <- flextable::align(table, align = "center", part = "all")
     table <- flextable::font(table, part = "all", fontname = "Times New Roman")
@@ -74,6 +66,7 @@ nice_fit <- function(..., nice_table = FALSE) {
 nice_fit_internal <- function(fit) {
   x <- lavaan::fitMeasures(fit, c("chisq", "df", "pvalue", "cfi", "tli",
                                   "rmsea", "srmr", "aic", "bic"))
+  #names(x)[1] <- "Model"
   x <- as.data.frame(t(as.data.frame(x)))
   chi2.df <- x$chisq / x$df
   x <- cbind(x[1:2], chi2.df, x[3:9])