+++ R/write_lavaan.R
@@ -2,26 +2,33 @@
 #'
 #' @description Vector-based lavaan syntax interpreter.
 #'
-#' @param mediation Mediation	indicators (`~` symbol: "is regressed on"). Differs from
-#'                  argument `regression` because path names can be optionally
-#'                  specified automatically with argument `label`.
+#' @param mediation Mediation	indicators (`~` symbol: "is regressed on").
+#'                  Differs from argument `regression` because path names
+#'                  can be optionally specified automatically with argument
+#'                  `label`.
 #' @param regression Regression indicators (`~` symbol: "is regressed on").
-#' @param covariance (Residual) (co)variance indicators (`~~` symbol: "is correlated with").
-#' @param indirect Indirect effect indicators (`:=` symbol: "indirect effect defined as").
-#'                 If a named list is provided, with names "IV" (independent variables), "M"
-#'                 (mediator), and "DV" (dependent variables), `write_lavaan` attempts to
-#'                 write indirect effects automatically. In this case, the `mediation`
-#'                 argument must be specified too.
+#' @param covariance (Residual) (co)variance indicators (`~~` symbol:
+#'                   "is correlated with").
+#' @param indirect Indirect effect indicators (`:=` symbol: "indirect
+#'                 effect defined as"). If a named list is provided,
+#'                 with names "IV" (independent variables), "M" (mediator),
+#'                 and "DV" (dependent variables), `write_lavaan` attempts to
+#'                 write indirect effects automatically. In this case, the
+#'                 `mediation` argument must be specified too.
 #' @param latent Latent variable indicators (`=~` symbol: "is measured by").
 #' @param intercept Intercept indicators (`~ 1` symbol: "intercept").
 #' @param constraint.equal Equality indicators (`==` symbol).
 #' @param constraint.smaller Smaller than indicators (`<` symbol).
 #' @param constraint.larger Greater than indicators (`>` symbol).
-#' @param custom Custom specifications. Takes a *single* string just like regular
-#'               `lavaan` syntax would. Always added at the end of the model.
-#' @param label Logical, whether to display path names for the mediation argument.
-#' @param use.letters Logical, for the labels, whether to use letters instead of the variable names.
-#' @keywords lavaan, structural equation modeling, path analysis, CFA
+#' @param custom Custom specifications. Takes a *single* string just
+#'               like regular `lavaan` syntax would. Always added at
+#'               the end of the model.
+#' @param label Logical, whether to display path names for the
+#'              mediation argument.
+#' @param use.letters Logical, for the labels, whether to use letters
+#'                    instead of the variable names.
+#' @keywords lavaan structural equation modeling path analysis CFA
+#' @return A character string, representing the specified `lavaan` model.
 #' @export
 #' @examples
 #' (latent <- list(visual = paste0("x", 1:3),
@@ -44,9 +51,9 @@ write_lavaan <- function(mediation = NULL, regression = NULL, covariance = NULL,
                          use.letters = FALSE) {
   constraint <- NULL
   hashtag <- sprintf("%s\n", paste0(rep("#", 50), collapse = ""))
-  process_vars <- function(x, symbol, label = FALSE, collapse = " + ",
-                           header = paste0(hashtag, paste0("# ", title, "\n\n")),
-                           title = NULL, spacer = "\n\n") {
+  process_vars <- function(
+    x, symbol, label = FALSE, collapse = " + ", header = paste0(
+      hashtag, paste0("# ", title, "\n\n")), title = NULL, spacer = "\n\n") {
     if(isTRUE(label)) {
       labels <- paste0(names(x))
       if (isTRUE(use.letters)) {
@@ -97,30 +104,34 @@ write_lavaan <- function(mediation = NULL, regression = NULL, covariance = NULL,
       stats::setNames(indirect.list, indirect.names)
       indirect <- stats::setNames(indirect.list, indirect.names)
     }
-    indirect <- process_vars(indirect, symbol = ":=", collapse = " * ", title =
-                               "[--------Mediations (indirect effects)---------]")
+    indirect <- process_vars(
+      indirect, symbol = ":=", collapse = " * ", title =
+        "[--------Mediations (indirect effects)---------]")
   }
   if (!is.null(mediation)) {
-    mediation <- process_vars(mediation, symbol = "~", label = label, title =
-                                "[-----------Mediations (named paths)-----------]")
+    mediation <- process_vars(
+      mediation, symbol = "~", label = label, title =
+        "[-----------Mediations (named paths)-----------]")
   }
   if (!is.null(regression)) {
-    regression <- process_vars(regression, symbol = "~", title =
-                                 "[---------Regressions (Direct effects)---------]")
+    regression <- process_vars(
+      regression, symbol = "~", title =
+        "[---------Regressions (Direct effects)---------]")
   }
   if (!is.null(covariance)) {
-    covariance <- process_vars(covariance, symbol = "~~", title =
-                                 "[------------------Covariances-----------------]")
+    covariance <- process_vars(
+      covariance, symbol = "~~", title =
+        "[------------------Covariances-----------------]")
   }
   if (!is.null(intercept)) {
-    title = "[------------------Intercepts------------------]"
+    title <- "[------------------Intercepts------------------]"
     header <- paste0(hashtag, paste0("# ", title, "\n\n"))
     intercept <- paste0(header, paste0(intercept, " ~ 1", collapse = "\n"),
                         "\n\n")
   }
   if (!is.null(constraint.equal) || !is.null(constraint.smaller) ||
       !is.null(constraint.larger)) {
-    title = "[-----------------Constraints------------------]"
+    title <- "[-----------------Constraints------------------]"
     header <- paste0(hashtag, paste0("# ", title, "\n\n"))
     if (!is.null(constraint.equal)) {
       constraint.equal <- process_vars(constraint.equal, symbol = "==",
@@ -138,7 +149,7 @@ write_lavaan <- function(mediation = NULL, regression = NULL, covariance = NULL,
                          constraint.larger, "\n")
   }
   if (!is.null(custom)) {
-    title = "[------------Custom Specifications-------------]"
+    title <- "[------------Custom Specifications-------------]"
     header <- paste0(hashtag, paste0("# ", title, "\n\n"))
     custom <- paste0(header, custom)
   }