+++ R/zzz.R
@@ -1,6 +1,7 @@
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage(
-    "Suggested citation: Th\u00e9riault, R. (2022). lavaanExtra: Convenience functions",
+    "Suggested citation: Th\u00e9riault, R. (2022).",
+    "lavaanExtra: Convenience functions",
     " for lavaan \n(R package version ",
     utils::packageVersion("lavaanExtra"),
     ") [Computer software]. https://lavaanExtra.remi-theriault.com/ "