+++ tests/testthat/test-save_as_x.R
@@ -1,0 +1,42 @@
+setwd(tempdir())
+
+latent <- list(visual = paste0("x", 1:3),
+               textual = paste0("x", 4:6),
+               speed = paste0("x", 7:9))
+
+regression <- list(ageyr = c("visual", "textual", "speed"),
+                   grade = c("visual", "textual", "speed"))
+
+HS.model <- write_lavaan(latent = latent, regression = regression)
+
+library(lavaan)
+fit <- sem(HS.model, data = HolzingerSwineford1939)
+nice.table <- nice_fit(fit, nice_table = TRUE)
+
+#   ____________________________________________________________________________
+#   Tests                                                                   ####
+
+
+test_that("nice_fit save for Word", {
+  expect_silent(
+    save_as_docx(nice.table, path = "table.docx")
+  )
+})
+
+test_that("nice_fit save for PPT", {
+  expect_silent(
+    save_as_pptx(nice.table, path = "table.pptx")
+  )
+})
+
+test_that("nice_fit save for html", {
+  expect_silent(
+    save_as_html(nice.table, path = "table.html")
+  )
+})
+
+test_that("nice_fit save for image", {
+  expect_silent(
+    save_as_image(nice.table, path = "table.png")
+  )
+})