+++ NEWS.md
@@ -1,3 +1,8 @@
+## lavaanExtra 0.1.2
+* `rempsyc` package dependency is now on CRAN, so we can rely on it normally now (yeah!).
+* Adding test coverage (100% so far!).
+* Preparing for CRAN.
+
 ## lavaanExtra 0.1.1
 * `nice_tidySEM`: Gains the `label_location` and `...` arguments.
 
@@ -24,7 +29,7 @@
 * `cfa_fit_plot`: 
   * Adds ellipsis to pass any desired argument to `lavaan::cfa`, and thus removes the default `missing = "fiml"` which was causing a bug with `estimator = DWLS`.
   * Changes the default from saving to PDF to opening in RStudio Viewer, and adds an additional `save.as.pdf` argument to save as PDF.
-  * Now imports `flextable::save_as_docx` (and `save_as_html`, `save_as_image`, `save_as_pptx`...) to support saving the `nice_fit` tables to Word without having to explicitely load `flextable` or `rempsyc`.
+  * Now imports `flextable::save_as_docx` (and `save_as_html`, `save_as_image`, `save_as_pptx`...) to support saving the `nice_fit` tables to Word without having to explicitly load `flextable` or `rempsyc`.
   * Added package logo!
 
 ## lavaanExtra 0.0.3